% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom.JSON}
\alias{custom.JSON}
\title{Generate JSON from a (manually prepared) table of features for your protein.}
\usage{
custom.JSON(
  protein.length = NULL,
  accession.number = NULL,
  description = NULL,
  organism = NULL,
  link.url = NULL,
  taxid = NULL,
  input.file = NULL,
  input.format = NULL,
  gradient = NULL
)
}
\arguments{
\item{protein.length}{integer. Protein length as a number of amino acids.}

\item{accession.number}{accession number (optional).}

\item{description}{short description/name of your protein}

\item{organism}{organism your protein is from}

\item{link.url}{web-link to online resource}

\item{taxid}{organism taxa ID}

\item{input.file}{input file (curated table with features)}

\item{input.format}{format of the table: XLSX, CSV, TSV}
}
\value{
Generates topology organisation scheme file in Pfam's JSON format.
}
\description{
Process a table of features of your protein of interest. Typically, that's a
manually curated list of features, summarized in a spreadsheet, which can be
imported. Generates topology organisation scheme file in Pfam's JSON format.
}
\examples{
library("jsonlite")
library("seqinr")
library("dplyr")
library("openxlsx")
library("gggenomes")
library("IRanges")
library("rentrez")
library("viridis")
library("protodeviser")

# path to CD45 examples
examples <- system.file("webApp/www/cd45/", package = "protodeviser")

# custom table file
custom   <- paste0(examples, "cd45_custom.xlsx")

cd45_hsap <- custom.JSON(protein.length = 1306,
               accession.number = "P08575",
               description = "CD45_HSAP",
               organism = "Homo_sapiens",
               link.url = "https://www.uniprot.org/uniprotkb/P08575/",
               taxid = "9606",
               input.file = custom,
               input.format = "xlsx",
	       gradient = "rainbow")
}
