% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id.JSON}
\alias{id.JSON}
\title{Generate JSON for a protein database identifier (ID).}
\usage{
id.JSON(
  input = NULL,
  database = NULL,
  offline = FALSE,
  regions.on = TRUE,
  motifs.on = TRUE,
  markups.on = TRUE,
  gradient = NULL
)
}
\arguments{
\item{input}{database accession number (identifier)}

\item{database}{database to be searched: UniProt or NCBI GenPept.}

\item{offline}{whether to search the database directly online (default) or use a local file.}

\item{regions.on}{shall we include regions in the results (default is TRUE).}

\item{motifs.on}{shall we include motifs in the results (default is TRUE).}

\item{markups.on}{shall we include markups in the results (default is TRUE).}
}
\value{
Generates topology organisation scheme file in Pfam's JSON format.
}
\description{
Search database (UniProt or NCBI GenPept) with submitted protein identifier
(database accession number) and extract protein features from the respective
database. Generates topology organisation scheme file in Pfam's JSON format.
}
\examples{
library("jsonlite")
library("seqinr")
library("dplyr")
library("openxlsx")
library("gggenomes")
library("IRanges")
library("rentrez")
library("viridis")
library("protodeviser")

# Scan from UniProt
P08575 <- id.JSON(input = "P08575", database = "uniprot", gradient = "rainbow")

# Scan from NCBI GenPept
NP_002829.3 <- id.JSON(input = "NP_002829.3", database = "ncbi", gradient = "rainbow")
}
