% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json.TABLE}
\alias{json.TABLE}
\title{Generate a table from JSON input of protein features.}
\usage{
json.TABLE(json = NULL)
}
\arguments{
\item{json}{JSON input to be written as a table}
}
\value{
A table generated from a JSON input.
}
\description{
Create a simple table from JSON (e.g. files saved on disk). Useful to compare
different annotations from different sources for the same protein. Only features
are exported, no metadata.
}
\examples{
library("jsonlite")
library("seqinr")
library("dplyr")
library("openxlsx")
library("gggenomes")
library("IRanges")
library("rentrez")
library("protodeviser")

# path to CD45 examples
examples <- system.file("webApp/www/cd45/", package = "protodeviser")

# CD45 JSON file
P08575 <- paste0(examples, "P08575.json")

# import using `fromJSON`
myjson <- fromJSON(P08575)

# convert to table
table <- json.TABLE(json = myjson)
}
