% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predicted.JSON}
\alias{predicted.JSON}
\title{Generate JSON from predicted features for your protein.}
\usage{
predicted.JSON(
  protein.length = NULL,
  accession.number = NULL,
  description = NULL,
  organism = NULL,
  link.url = NULL,
  taxid = NULL,
  SMART.tsv = NULL,
  ELM.xlsx = NULL,
  ELM.features.tsv = NULL,
  ELM.score = NULL,
  SCANSITE.tsv = NULL,
  SCANSITE.score = NULL,
  SCANSITE.percentile = NULL,
  SCANSITE.accessibility = NULL,
  ANCHOR.tsv = NULL,
  ANCHOR.cutoff = NULL,
  netNglyc.tsv = NULL,
  netNglyc.cutoff = NULL,
  netOglyc.tsv = NULL,
  netOglyc.cutoff = NULL,
  netPhos.tsv = NULL,
  netPhos.cutoff = NULL,
  gradient = NULL
)
}
\arguments{
\item{protein.length}{integer. Protein length as a number of amino acids.}

\item{accession.number}{accession number (optional).}

\item{description}{short description/name of your protein}

\item{organism}{organism your protein is from}

\item{link.url}{web-link to online resource}

\item{taxid}{organism taxa ID}

\item{SMART.tsv}{SMART database results file}

\item{ELM.xlsx}{ELM database results file}

\item{ELM.features.tsv}{ELM conservation scores file}

\item{ELM.score}{ELM conservation scores cutoff}

\item{SCANSITE.tsv}{ScanSite results file}

\item{SCANSITE.score}{ScanSite scores cutoff}

\item{SCANSITE.percentile}{ScanSite percentile cutoff}

\item{SCANSITE.accessibility}{ScanSite accessibility cutoff}

\item{ANCHOR.tsv}{Anchor2 results file}

\item{ANCHOR.cutoff}{Anchor2 scores cutoff}

\item{netNglyc.tsv}{NetNGlyc results file}

\item{netNglyc.cutoff}{NetNGlyc scores cutoff}

\item{netOglyc.tsv}{NetOGlyc results file}

\item{netOglyc.cutoff}{NetOGlyc scores cutoff}

\item{netPhos.tsv}{NetPhos results file}

\item{netPhos.cutoff}{NetPhos scores cutoff}
}
\value{
Generates topology organisation scheme file in Pfam's JSON format.
}
\description{
Process protein features predicted at different online resources. These are:
SMART, ELM, IUPred3/Anchor2, NetNGlyc, NetOGlyc, NetPhos and ScanSites. Cut-
off values for scores can be specified. Metadata, such as protein length
(required) and (optionally) protein name and other info should be provided by
the user. Generates topology organisation scheme file in Pfam's JSON format.
}
\examples{
library("jsonlite")
library("seqinr")
library("dplyr")
library("openxlsx")
library("gggenomes")
library("IRanges")
library("rentrez")
library("viridis")
library("protodeviser")

# path to CD45 examples of predicted results
examples <- system.file("webApp/www/cd45/", package = "protodeviser")

# files to submit to function
cd45_SMART    <- paste0(examples, "cd45_SMART.tsv")
cd45_ELM      <- paste0(examples, "cd45_ELM.xlsx")
cd45_ELM_cs   <- paste0(examples, "cd45_ELMcs.tsv")
cd45_SCANSITE <- paste0(examples, "cd45_ScanSite.tsv")
cd45_ANCHOR   <- paste0(examples, "cd45_IUPred3.tsv")
cd45_netNglyc <- paste0(examples, "cd45_NetNGlyc.tsv")
cd45_netOglyc <- paste0(examples, "cd45_NetOGlyc.gff")
cd45_netPhos  <- paste0(examples, "cd45_NetPhos.tsv")

cd45_predicted <- predicted.JSON(protein.length = 1306,
               accession.number = "P08575",
               description = "CD45_HSAP",
               organism = "Homo_sapiens",
               link.url = "https://www.uniprot.org/uniprotkb/P08575/",
               taxid = "9606",
               SMART.tsv = cd45_SMART,
               ELM.xlsx = cd45_ELM,
               ELM.features.tsv = cd45_ELM_cs,
               ELM.score = 0.5,
               SCANSITE.tsv = cd45_SCANSITE,
               SCANSITE.score = 0.5,
               SCANSITE.percentile = 0.001,
               SCANSITE.accessibility = 1,
               ANCHOR.tsv = cd45_ANCHOR,
               ANCHOR.cutoff = 0.8,
               netNglyc.tsv = cd45_netNglyc,
               netNglyc.cutoff = 0.9,
               netOglyc.tsv = cd45_netOglyc,
               netOglyc.cutoff = 0.9,
               netPhos.tsv = cd45_netPhos,
               netPhos.cutoff = 0.95,
	       gradient = "rainbow")
}
