% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protodeviser_ui.R
\name{protodeviser_ui}
\alias{protodeviser_ui}
\title{ProToDeviseR UI}
\usage{
protodeviser_ui()
}
\value{
A graphical user interface for ProToDeviser
}
\description{
ProToDeviseR main app with graphical user interface (UI)
}
